// tests/protoc-gen-ts-grpc-gateway-cli/integration/defaultConfig/proto/examplepb/a_bit_of_everything.ts
export interface ABitOfEverything {
  singleNested?: ABitOfEverything_Nested | undefined;
  uuid: string;
  nested: ABitOfEverything_Nested[];
  floatValue: number;
  doubleValue: number;
  int64Value: number;
  uint64Value: number;
  int32Value: number;
  fixed64Value: number;
  fixed32Value: number;
  boolValue: boolean;
  stringValue: string;
  bytesValue: Uint8Array;
  uint32Value: number;
  enumValue: NumericEnum;
  pathEnumValue: PathEnum;
  nestedPathEnumValue: MessagePathEnum_NestedPathEnum;
  sfixed32Value: number;
  sfixed64Value: number;
  sint32Value: number;
  sint64Value: number;
  repeatedStringValue: string[];
  oneofEmpty?: Empty | undefined;
  oneofString?: string | undefined;
  mapValue: { [key: string]: NumericEnum };
  mappedStringValue: { [key: string]: string };
  mappedNestedValue: { [key: string]: ABitOfEverything_Nested };
  timestampValue?: Date | undefined;
  /** repeated enum value. it is comma-separated in query */
  repeatedEnumValue: NumericEnum[];
  /** repeated numeric enum comment (This comment is overridden by the field annotation) */
  repeatedEnumAnnotation: NumericEnum[];
  /** numeric enum comment (This comment is overridden by the field annotation) */
  enumValueAnnotation: NumericEnum;
  /** repeated string comment (This comment is overridden by the field annotation) */
  repeatedStringAnnotation: string[];
  /** repeated nested object comment (This comment is overridden by the field annotation) */
  repeatedNestedAnnotation: ABitOfEverything_Nested[];
  /** nested object comments (This comment is overridden by the field annotation) */
  nestedAnnotation?: ABitOfEverything_Nested | undefined;
  int64OverrideType: number;
  /** mark a field as required in Open API definition */
  requiredStringViaFieldBehaviorAnnotation: string;
  /** mark a field as readonly in Open API definition */
  outputOnlyStringViaFieldBehaviorAnnotation: string;
  optionalStringValue?: string | undefined;
  /** Test openapiv2 generation of repeated fields */
  productId: string[];
  /** Test openapiv2 generation of required fields with annotation and jsonschema to reproduce */
  optionalStringField: string;
  requiredStringField1: string;
  requiredStringField2: string;
  /** Test openapiv2 handling of required json_name fields */
  requiredFieldBehaviorJsonName: string;
  requiredFieldSchemaJsonName: string;
  /** Trailing only */
  trailingOnly: string;
  /** Trailing only dot. */
  trailingOnlyDot: string;
  /** Leading both */
  trailingBoth: string;
  /**
   * Leading multiline
   *
   * This is an example of a multi-line comment.
   */
  trailingMultiline: string;
  /** Specify a custom format of repeated field items */
  uuids: string[];
}

/** Nested is nested type. */
export interface ABitOfEverything_Nested {
  /** name is nested field. */
  name: string;
  amount: number;
  /** DeepEnum comment. */
  ok: ABitOfEverything_Nested_DeepEnum;
}
// ...
export interface ABitOfEverythingRepeated {
  /** repeated values. they are comma-separated in path */
  pathRepeatedFloatValue: number[];
  pathRepeatedDoubleValue: number[];
  pathRepeatedInt64Value: number[];
  pathRepeatedUint64Value: number[];
  pathRepeatedInt32Value: number[];
  pathRepeatedFixed64Value: number[];
  pathRepeatedFixed32Value: number[];
  pathRepeatedBoolValue: boolean[];
  pathRepeatedStringValue: string[];
  pathRepeatedBytesValue: Uint8Array[];
  pathRepeatedUint32Value: number[];
  pathRepeatedEnumValue: NumericEnum[];
  pathRepeatedSfixed32Value: number[];
  pathRepeatedSfixed64Value: number[];
  pathRepeatedSint32Value: number[];
  pathRepeatedSint64Value: number[];
}

export interface CheckStatusResponse {
  status?: Status | undefined;
}

export interface Body {
  name: string;
}
// ...
export const ABitOfEverything: MessageFns<ABitOfEverything> = {
  fromJSON(object: any): ABitOfEverything {
    return {
      singleNested: isSet(object.singleNested)
        ? ABitOfEverything_Nested.fromJSON(object.singleNested)
        : undefined,
      uuid: isSet(object.uuid) ? globalThis.String(object.uuid) : "",
      nested: globalThis.Array.isArray(object?.nested)
        ? object.nested.map((e: any) => ABitOfEverything_Nested.fromJSON(e))
        : [],
      floatValue: isSet(object.floatValue)
        ? globalThis.Number(object.floatValue)
        : 0,
      doubleValue: isSet(object.doubleValue)
        ? globalThis.Number(object.doubleValue)
        : 0,
      int64Value: isSet(object.int64Value)
        ? globalThis.Number(object.int64Value)
        : 0,
      uint64Value: isSet(object.uint64Value)
        ? globalThis.Number(object.uint64Value)
        : 0,
      int32Value: isSet(object.int32Value)
        ? globalThis.Number(object.int32Value)
        : 0,
      fixed64Value: isSet(object.fixed64Value)
        ? globalThis.Number(object.fixed64Value)
        : 0,
      fixed32Value: isSet(object.fixed32Value)
        ? globalThis.Number(object.fixed32Value)
        : 0,
      boolValue: isSet(object.boolValue)
        ? globalThis.Boolean(object.boolValue)
        : false,
      stringValue: isSet(object.stringValue)
        ? globalThis.String(object.stringValue)
        : "",
      bytesValue: isSet(object.bytesValue)
        ? bytesFromBase64(object.bytesValue)
        : new Uint8Array(0),
      uint32Value: isSet(object.uint32Value)
        ? globalThis.Number(object.uint32Value)
        : 0,
      enumValue: isSet(object.enumValue)
        ? numericEnumFromJSON(object.enumValue)
        : NumericEnum.ZERO,
      pathEnumValue: isSet(object.pathEnumValue)
        ? pathEnumFromJSON(object.pathEnumValue)
        : PathEnum.ABC,
      nestedPathEnumValue: isSet(object.nestedPathEnumValue)
        ? messagePathEnum_NestedPathEnumFromJSON(object.nestedPathEnumValue)
        : MessagePathEnum_NestedPathEnum.GHI,
      sfixed32Value: isSet(object.sfixed32Value)
        ? globalThis.Number(object.sfixed32Value)
        : 0,
      sfixed64Value: isSet(object.sfixed64Value)
        ? globalThis.Number(object.sfixed64Value)
        : 0,
      sint32Value: isSet(object.sint32Value)
        ? globalThis.Number(object.sint32Value)
        : 0,
      sint64Value: isSet(object.sint64Value)
        ? globalThis.Number(object.sint64Value)
        : 0,
      repeatedStringValue: globalThis.Array.isArray(object?.repeatedStringValue)
        ? object.repeatedStringValue.map((e: any) => globalThis.String(e))
        : [],
      oneofEmpty: isSet(object.oneofEmpty)
        ? Empty.fromJSON(object.oneofEmpty)
        : undefined,
      oneofString: isSet(object.oneofString)
        ? globalThis.String(object.oneofString)
        : undefined,
      mapValue: isObject(object.mapValue)
        ? Object.entries(object.mapValue).reduce<{
            [key: string]: NumericEnum;
          }>((acc, [key, value]) => {
            acc[key] = numericEnumFromJSON(value);
            return acc;
          }, {})
        : {},
// ...
    };
  },
// ...
};

function createBaseABitOfEverything_Nested(): ABitOfEverything_Nested {
  return { name: "", amount: 0, ok: ABitOfEverything_Nested_DeepEnum.FALSE };
}

export const ABitOfEverything_Nested: MessageFns<ABitOfEverything_Nested> = {
  fromJSON(object: any): ABitOfEverything_Nested {
    return {
      name: isSet(object.name) ? globalThis.String(object.name) : "",
      amount: isSet(object.amount) ? globalThis.Number(object.amount) : 0,
      ok: isSet(object.ok)
        ? aBitOfEverything_Nested_DeepEnumFromJSON(object.ok)
        : ABitOfEverything_Nested_DeepEnum.FALSE,
    };
  },

  toJSON(message: ABitOfEverything_Nested): unknown {
    const obj: any = {};
    if (message.name !== "") {
      obj.name = message.name;
    }
    if (message.amount !== 0) {
      obj.amount = Math.round(message.amount);
    }
    if (message.ok !== ABitOfEverything_Nested_DeepEnum.FALSE) {
      obj.ok = aBitOfEverything_Nested_DeepEnumToJSON(message.ok);
    }
    return obj;
  },

  create(base?: DeepPartial<ABitOfEverything_Nested>): ABitOfEverything_Nested {
    return ABitOfEverything_Nested.fromPartial(base ?? {});
  },
  fromPartial(
    object: DeepPartial<ABitOfEverything_Nested>,
  ): ABitOfEverything_Nested {
    const message = createBaseABitOfEverything_Nested();
    message.name = object.name ?? "";
    message.amount = object.amount ?? 0;
    message.ok = object.ok ?? ABitOfEverything_Nested_DeepEnum.FALSE;
    return message;
  },
};
// ...
export const Body: MessageFns<Body> = {
  fromJSON(object: any): Body {
    return { name: isSet(object.name) ? globalThis.String(object.name) : "" };
  },

  toJSON(message: Body): unknown {
    const obj: any = {};
    if (message.name !== "") {
      obj.name = message.name;
    }
    return obj;
  },

  create(base?: DeepPartial<Body>): Body {
    return Body.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<Body>): Body {
    const message = createBaseBody();
    message.name = object.name ?? "";
    return message;
  },
};
// ...
export interface ABitOfEverythingServiceClient<CallOptionsExt = {}> {
  createBody(
    request: DeepPartial<ABitOfEverything>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<ABitOfEverything>;
  lookup(
    request: DeepPartial<IdMessage>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<ABitOfEverything>;
  custom(
    request: DeepPartial<ABitOfEverything>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<ABitOfEverything>;
  doubleColon(
    request: DeepPartial<ABitOfEverything>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<ABitOfEverything>;
  update(
    request: DeepPartial<ABitOfEverything>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<Empty>;
  updateV2(
    request: DeepPartial<UpdateV2Request>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<Empty>;
  delete(
    request: DeepPartial<IdMessage>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<Empty>;
  getQuery(
    request: DeepPartial<ABitOfEverything>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<Empty>;
  getRepeatedQuery(
    request: DeepPartial<ABitOfEverythingRepeated>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<ABitOfEverythingRepeated>;
  deepPathEcho(
    request: DeepPartial<ABitOfEverything>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<ABitOfEverything>;
  noBindings(
    request: DeepPartial<Duration>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<Empty>;
  timeout(
    request: DeepPartial<Empty>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<Empty>;
  errorWithDetails(
    request: DeepPartial<Empty>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<Empty>;
  getMessageWithBody(
    request: DeepPartial<MessageWithBody>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<Empty>;
  postWithEmptyBody(
    request: DeepPartial<Body>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<Empty>;
  checkGetQueryParams(
    request: DeepPartial<ABitOfEverything>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<ABitOfEverything>;
  checkNestedEnumGetQueryParams(
    request: DeepPartial<ABitOfEverything>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<ABitOfEverything>;
  checkPostQueryParams(
    request: DeepPartial<ABitOfEverything>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<ABitOfEverything>;
  overwriteRequestContentType(
    request: DeepPartial<Body>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<Empty>;
  overwriteResponseContentType(
    request: DeepPartial<Empty>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<StringValue>;
  checkExternalPathEnum(
    request: DeepPartial<MessageWithPathEnum>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<Empty>;
  checkExternalNestedPathEnum(
    request: DeepPartial<MessageWithNestedPathEnum>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<Empty>;
  checkStatus(
    request: DeepPartial<Empty>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<CheckStatusResponse>;
  postOneofEnum(
    request: DeepPartial<OneofEnumMessage>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<Empty>;
  postRequiredMessageType(
    request: DeepPartial<RequiredMessageTypeRequest>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<Empty>;
}
// ...
export interface AnotherServiceWithNoBindingsClient<CallOptionsExt = {}> {
  noBindings(
    request: DeepPartial<Empty>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<Empty>;
}